+++ server/repository/settings-repository.go
@@ -42,6 +42,7 @@ var (
 	SettingMaxHoursBeforeDelete           SettingName = SettingName{Name: "max_hours_before_delete", Type: SettingTypeInt}
 	SettingMinBookingDurationHours        SettingName = SettingName{Name: "min_booking_duration_hours", Type: SettingTypeInt}
 	SettingMaxBookingDurationHours        SettingName = SettingName{Name: "max_booking_duration_hours", Type: SettingTypeInt}
+	SettingTargetUtilizationHoursPerWeek  SettingName = SettingName{Name: "target_utilization_hours_per_week", Type: SettingTypeInt}
 	SettingMaxHoursPartiallyBooked        SettingName = SettingName{Name: "max_hours_partially_booked", Type: SettingTypeInt}
 	SettingMaxHoursPartiallyBookedEnabled SettingName = SettingName{Name: "max_hours_partially_booked_enabled", Type: SettingTypeBool}
 	SettingDailyBasisBooking              SettingName = SettingName{Name: "daily_basis_booking", Type: SettingTypeBool}
@@ -274,6 +275,7 @@ func (r *SettingsRepository) InitDefaultSettingsForOrg(organizationID string) er
 		"($1, '"+SettingMaxHoursPartiallyBookedEnabled.Name+"', '0'), "+
 		"($1, '"+SettingMaxHoursPartiallyBooked.Name+"', '8'), "+
 		"($1, '"+SettingMinBookingDurationHours.Name+"', '0'), "+
+		"($1, '"+SettingTargetUtilizationHoursPerWeek.Name+"', '40'), "+
 		"($1, '"+SettingMaxDaysInAdvance.Name+"', '14'), "+
 		"($1, '"+SettingMaxBookingDurationHours.Name+"', '12'), "+
 		"($1, '"+SettingDefaultTimezone.Name+"', 'Europe/Berlin'), "+